Update Unit test and Username check for each end point
