+++ src/test/java/com/webapp/flightsearch/service/FlightSearchServiceTests.java
@@ -1,0 +1,5 @@
+package com.webapp.flightsearch.service;
+
+public class FlightSearchServiceTests {
+
+}
